requirements.txt
aerich==0.5.4
aiosqlite==0.16.1
asgiref==3.4.1
asyncpg==0.23.0
click==8.0.1
colorama==0.4.4
ddlparse==1.10.0
dictdiffer==0.8.1
fastapi==0.66.1
h11==0.12.0
httptools==0.2.0
iso8601==0.1.16
pydantic==1.8.2
pyparsing==2.4.7
pypika-tortoise==0.1.1
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
starlette==0.14.2
tortoise-orm==0.17.5
typing-extensions==3.10.0.0
uvicorn==0.14.0
watchgod==0.7
websockets==9.1
